move: move SubPageHeader and test to directory
