+++ src/shared/ui/SubPageHeader/SubPaggeHeader.test.tsx
@@ -1,7 +1,9 @@
 import { describe, it, expect } from 'vitest';
 import { render } from '@testing-library/react';
 
-import SubPageHeader, { SubPageHeaderProps } from '../SubPageHeader';
+import SubPageHeader, {
+  SubPageHeaderProps,
+} from '../SubPageHeader/SubPageHeader';
 
 describe('SubPageHeader', () => {
   const renderSubPageHeader = ({ backLink, title }: SubPageHeaderProps) =>
